requirements.txt
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.0
requests==2.25.1
torch==1.7.1
beautifulsoup4~=4.9.3
retry~=0.9.2
PyGithub==1.55
